Refactor to avoid unnecessary prop drilling
