+++ packages/strapi-design-system/src/Json/JsonInput.stories.mdx
@@ -36,14 +36,14 @@ JSON input allows users to add, edit or view json data in a dedicated editor.The
   <Story
     name="base"
     parameters={{
-      docs: { 
+      docs: {
         source: {
           code: `<JsonInput value={'[\n   {\n      "a":3,\n      "b":4\n   },\n   {\n      "a":5,\n      "b":6\n   }\n]'} />`,
         },
       },
     }}
   >
-      <JsonInput disabled value={'[\n   {\n      "a":3,\n      "b":4\n   },\n   {\n      "a":5,\n      "b":6\n   }\n]'} />
+    <JsonInput disabled value={'[\n   {\n      "a":3,\n      "b":4\n   },\n   {\n      "a":5,\n      "b":6\n   }\n]'} />
   </Story>
 </Canvas>
 
@@ -55,14 +55,18 @@ JSON input can be defined as read-only or can allow you to interact and edit the
   <Story
     name="editable"
     parameters={{
-      docs: { 
+      docs: {
         source: {
           code: `<JsonInput value={'[\n   {\n      "a":3,\n      "b":4\n   },\n   {\n      "a":5,\n      "b":6\n   }\n]'} label="Json" editable hint="Description Text" />`,
         },
       },
     }}
   >
-      <JsonInput value={'[\n   {\n      "a":3,\n      "b":4\n   },\n   {\n      "a":5,\n      "b":6\n   }\n]'} label="Json" hint="Description Text" />
+    <JsonInput
+      value={'[\n   {\n      "a":3,\n      "b":4\n   },\n   {\n      "a":5,\n      "b":6\n   }\n]'}
+      label="Json"
+      hint="Description Text"
+    />
   </Story>
 </Canvas>
 
@@ -74,14 +78,17 @@ JSON input can display an error. The error message is customizable and replaces 
   <Story
     name="withError"
     parameters={{
-      docs: { 
+      docs: {
         source: {
           code: `<JsonInput value={'[\n   {\n      "a":3,\n      "b":4\n   },\n   {\n      "a":5,\n      "b":6,\n   }\n]'} error="This does not match the JSON format"/>`,
         },
       },
     }}
   >
-     <JsonInput disabled value={'[\n   {\n      "a":3,\n      "b":4\n   },\n   {\n      "a":5,\n      "b":6,\n   }\n]'} error="This does not match the JSON format" />
+    <JsonInput
+      value={'[\n   {\n      "a":3,\n      "b":4\n   },\n   {\n      "a":5,\n      "b":6,\n   }\n]'}
+      error="This does not match the JSON format"
+    />
   </Story>
 </Canvas>
